rename structured log model module
