+++ logscale-client/Cargo.toml
@@ -13,6 +13,7 @@ repository = "https://github.com/sunero4/logscale-rs"
 [dependencies]
 reqwest = { version = "0.11.18", features = ["json"] }
 serde = { version = "1.0.182", features = ["derive"] }
+erased-serde = "0.3"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full"] }
