+++ logscale-client/examples/ingest_structured.rs
@@ -9,7 +9,7 @@ use std::{
 
 use logscale_client::{
     client::LogScaleClient,
-    models::structured_data::{StructuredLogEvent, StructuredLogsIngestRequest},
+    models::structured_logging::{StructuredLogEvent, StructuredLogsIngestRequest},
 };
 
 #[tokio::main]
