+++ logscale-client/src/models/structured_logging.rs
