+++ logscale-log/src/ingest_job.rs
@@ -7,7 +7,7 @@ use std::{
 
 use logscale_client::{
     client::LogScaleClient,
-    models::structured_data::{StructuredLogEvent, StructuredLogsIngestRequest},
+    models::structured_logging::{StructuredLogEvent, StructuredLogsIngestRequest},
 };
 
 use crate::log_events_cache::LogsEventCache;
