+++ logscale-log/src/log_events_cache.rs
@@ -1,4 +1,4 @@
-use logscale_client::models::structured_data::StructuredLogEvent;
+use logscale_client::models::structured_logging::StructuredLogEvent;
 
 pub struct LogsEventCache {
     cached_log_events: Vec<StructuredLogEvent>,
